requirements.txt
discordpy
database
datetime
settings
sqlalchemy